lyrics_parser_requirements.txt
altgraph==0.17.2
asttokens==2.0.8
auto-py-to-exe==2.23.0
backcall==0.2.0
bottle==0.12.23
bottle-websocket==0.2.9
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
cloudpickle==2.1.0
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
Eel==0.14.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
keyboard==0.13.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.4
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==23.2.1
six==1.16.0
SQLAlchemy==1.4.40
stack-data==0.5.0
tabulate==0.8.10
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
wcwidth==0.2.5
whichcraft==0.6.1
wincertstore==0.2
XlsxWriter==3.0.3
zope.event==4.5.0
zope.interface==5.4.0
